Refactor: getLexResponse의 결과를 BasicResponseMessageType[]으로 변경 + 파일 이동
