+++ src/store/message/fetchResponse.ts
@@ -1,14 +1,14 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import { ContentResponseMessageType } from "../../Interface/Message/ResponseMessageType";
+import { BasicResponseMessageType } from "../../Interface/Message/ResponseMessageType";
 import {
 	createResponse,
 	createResponseContent,
 } from "../../utils/Message/createResponse";
-import { getLexResponse } from "./getLexResponse";
 import {
-	defaultContentResponseMessageData,
 	defaultCardResponseMessageData,
-} from "./defaultResponseMessageData";
+	defaultContentResponseMessageData,
+} from "../../utils/Message/defaultResponseMessageData";
+import { getLexResponse } from "../../utils/Message/getLexResponse";
 
 export const fetchResponse = createAsyncThunk(
 	"message/fetchResponse",
@@ -32,7 +32,7 @@ export const fetchResponse = createAsyncThunk(
 );
 
 const isErrorMessage = (
-	response: ContentResponseMessageType[]
+	response: BasicResponseMessageType[]
 ): string | null => {
 	if (Object.keys(response).filter((key) => key.includes("error")).length > 0)
 		return "답변 생성에 오류가 있었습니다. 잠시 후 다시 시도해주세요.";
